+++ app/src/AppContainer.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Platform } from 'react-native';
 import { createAppContainer, createSwitchNavigator, NavigationContainer } from 'react-navigation';
 
-import { ChangesScreen } from './screens/ChangesScreen';
-import { SplashScreen } from './screens/SplashScreen';
+import { ChangesScreen } from './screens/changes/ChangesScreen';
+import { SplashScreen } from './screens/splash/SplashScreen';
 
 export function createMyAppContainer(): NavigationContainer {
     const AppNavigator: NavigationContainer = createSwitchNavigator(
