+++ app/src/screens/splash/SplashScreen.tsx
